ACL & Cleaning

+++ xbless/app/Http/Controllers/DiskonController.php
@@ -104,7 +104,7 @@ class DiskonController extends Controller
         $start = $request->start;
         $page  = $start +1;
         $search = $request->search['value'];
-        
+
         $query = DiskonDetail::select('diskon_detail.*','tbl_product.nama as name_product', 'tbl_satuan.nama as satuan_name');
         $query->leftJoin('tbl_product', 'tbl_product.id', 'diskon_detail.produk');
         $query->leftJoin('tbl_satuan','tbl_satuan.id','diskon_detail.satuan');
@@ -120,9 +120,9 @@ class DiskonController extends Controller
           });
         }
         $totalData = $query->get()->count();
-  
+
         $totalFiltered = $query->get()->count();
-  
+
         $query->limit($limit);
         $query->offset($start);
         $data = $query->get();
@@ -130,11 +130,15 @@ class DiskonController extends Controller
         {
             $enc_id = $this->safe_encode(Crypt::encryptString($result->id));
             $action = "";
-            
+
             $action.="";
             $action.="<div class='btn-group'>";
-            $action.='<a href="'.route('diskon.ubah',$enc_id).'" class="btn btn-warning btn-xs icon-btn md-btn-flat product-tooltip" title="Edit"><i class="fa fa-pencil"></i> Edit</a>&nbsp;';
-            $action.='<a href="#" onclick="deleteNegara(this,\''.$enc_id.'\')" class="btn btn-danger btn-xs icon-btn md-btn-flat product-tooltip" title="Hapus"><i class="fa fa-times"></i> Hapus</a>&nbsp;';
+            if($request->user()->can('diskon.ubah')) {
+                $action.='<a href="'.route('diskon.ubah',$enc_id).'" class="btn btn-warning btn-xs icon-btn md-btn-flat product-tooltip" title="Edit"><i class="fa fa-pencil"></i> Edit</a>&nbsp;';
+            }
+            if($request->user()->can('diskon.delete')) {
+                $action.='<a href="#" onclick="deleteNegara(this,\''.$enc_id.'\')" class="btn btn-danger btn-xs icon-btn md-btn-flat product-tooltip" title="Hapus"><i class="fa fa-times"></i> Hapus</a>&nbsp;';
+            }
             $action.="</div>";
 
             if($result->flag_diskon == 0){
@@ -166,30 +170,23 @@ class DiskonController extends Controller
             $result->keterangan     = $keterangan;
             $result->action         = $action;
         }
-  
-        $json_data = array(
-          "draw"            => intval($request->input('draw')),  
-          "recordsTotal"    => intval($totalData),
-          "recordsFiltered" => intval($totalFiltered),
-          "data"            => $data
-        );
-        // if($request->user()->can('negara.index')) {
-        //   $json_data = array(
-        //     "draw"            => intval($request->input('draw')),  
-        //     "recordsTotal"    => intval($totalData),
-        //     "recordsFiltered" => intval($totalFiltered),
-        //     "data"            => $data
-        //   );
-        // }else{
-        //   $json_data = array(
-        //     "draw"            => intval($request->input('draw')),  
-        //     "recordsTotal"    => 0,
-        //     "recordsFiltered" => 0,
-        //     "data"            => []
-        //   );
-           
-        // }    
-        return json_encode($json_data); 
+
+        if($request->user()->can('diskon.index')) {
+            $json_data = array(
+                "draw"            => intval($request->input('draw')),
+                "recordsTotal"    => intval($totalData),
+                "recordsFiltered" => intval($totalFiltered),
+                "data"            => $data
+            );
+        }else{
+            $json_data = array(
+                "draw"            => intval($request->input('draw')),
+                "recordsTotal"    => 0,
+                "recordsFiltered" => 0,
+                "data"            => []
+            );
+        }
+        return json_encode($json_data);
       }
 
     public function simpan(Request $req){
@@ -229,7 +226,7 @@ class DiskonController extends Controller
                 $json_data = array(
                     "success"         => FALSE,
                     "message"         => 'Data gagal ditambahkan.'
-                ); 
+                );
             }
         }else{
             $detail     = new DiskonDetail;
@@ -249,7 +246,7 @@ class DiskonController extends Controller
             $detail->save();
 
             if($detail){
-                
+
                 $json_data = array(
                     "success"         => TRUE,
                     "message"         => 'Data berhasil ditambahkan.'
@@ -259,11 +256,11 @@ class DiskonController extends Controller
                 $json_data = array(
                     "success"         => FALSE,
                     "message"         => 'Data gagal ditambahkan.'
-                ); 
+                );
             }
         }
 
-        return json_encode($json_data); 
+        return json_encode($json_data);
     }
 
     public function ubah($enc_id){
