+++ xbless/app/Http/Controllers/DistrikController.php
@@ -39,7 +39,7 @@ class DistrikController extends Controller
       $start = $request->start;
       $page  = $start +1;
       $search = $request->search['value'];
-      
+
       $country = Distrik::select('*');
       if(array_key_exists($request->order[0]['column'], $this->original_column)){
          $country->orderByRaw($this->original_column[$request->order[0]['column']].' '.$request->order[0]['dir']);
@@ -62,11 +62,15 @@ class DistrikController extends Controller
       {
         $enc_id = $this->safe_encode(Crypt::encryptString($negara->id));
         $action = "";
-       
+
         $action.="";
         $action.="<div class='btn-group'>";
-        $action.='<a href="'.route('distrik.ubah',$enc_id).'" class="btn btn-warning btn-xs icon-btn md-btn-flat product-tooltip" title="Edit"><i class="fa fa-pencil"></i> Edit</a>&nbsp;';
-        $action.='<a href="#" onclick="deleteNegara(this,\''.$enc_id.'\')" class="btn btn-danger btn-xs icon-btn md-btn-flat product-tooltip" title="Hapus"><i class="fa fa-times"></i> Hapus</a>&nbsp;';
+        if($request->user()->can('distrik.ubah')){
+            $action.='<a href="'.route('distrik.ubah',$enc_id).'" class="btn btn-warning btn-xs icon-btn md-btn-flat product-tooltip" title="Edit"><i class="fa fa-pencil"></i> Edit</a>&nbsp;';
+        }
+        if($request->user()->can('distrik.delete')){
+            $action.='<a href="#" onclick="deleteNegara(this,\''.$enc_id.'\')" class="btn btn-danger btn-xs icon-btn md-btn-flat product-tooltip" title="Hapus"><i class="fa fa-times"></i> Hapus</a>&nbsp;';
+        }
         $action.="</div>";
 
         $negara->no             = $key+$page;
@@ -75,29 +79,23 @@ class DistrikController extends Controller
         $negara->action         = $action;
       }
 
-      $json_data = array(
-        "draw"            => intval($request->input('draw')),  
-        "recordsTotal"    => intval($totalData),
-        "recordsFiltered" => intval($totalFiltered),
-        "data"            => $data
-      );
-      // if($request->user()->can('negara.index')) {
-      //   $json_data = array(
-      //     "draw"            => intval($request->input('draw')),  
-      //     "recordsTotal"    => intval($totalData),
-      //     "recordsFiltered" => intval($totalFiltered),
-      //     "data"            => $data
-      //   );
-      // }else{
-      //   $json_data = array(
-      //     "draw"            => intval($request->input('draw')),  
-      //     "recordsTotal"    => 0,
-      //     "recordsFiltered" => 0,
-      //     "data"            => []
-      //   );
-         
-      // }    
-      return json_encode($json_data); 
+
+        if($request->user()->can('distrik.index')) {
+            $json_data = array(
+            "draw"            => intval($request->input('draw')),
+            "recordsTotal"    => intval($totalData),
+            "recordsFiltered" => intval($totalFiltered),
+            "data"            => $data
+            );
+        }else{
+            $json_data = array(
+            "draw"            => intval($request->input('draw')),
+            "recordsTotal"    => 0,
+            "recordsFiltered" => 0,
+            "data"            => []
+            );
+        }
+        return json_encode($json_data);
     }
 
     public function tambah(){
@@ -111,7 +109,7 @@ class DistrikController extends Controller
           $dec_id = $this->safe_decode(Crypt::decryptString($enc_id));
         }else{
           $dec_id = null;
-        }   
+        }
         $cek_negara = $this->cekExist('name',$req->name,$dec_id);
         if(!$cek_negara){
             $json_data = array(
@@ -148,11 +146,11 @@ class DistrikController extends Controller
                     "success"         => FALSE,
                     "message"         => 'Data gagal ditambahkan.'
               );
-            }   
+            }
           }
         }
 
-        return json_encode($json_data); 
+        return json_encode($json_data);
     }
 
     public function ubah($enc_id){
