+++ xbless/app/Http/Controllers/FinanceController.php
@@ -16,7 +16,7 @@ class FinanceController extends Controller
 {
     protected $original_column = array(
         1 => "name",
-    );   
+    );
 
     public function index(){
         return view('backend/pembayaran/finance/index');
@@ -26,12 +26,12 @@ class FinanceController extends Controller
         $data = str_replace(array('/'),array('_'),$string);
         return $data;
     }
-  
+
     function safe_decode($string,$mode=null) {
         $data = str_replace(array('_'),array('/'),$string);
         return $data;
     }
-  
+
     private function cekExist($column,$var,$id){
         $cek = Finance::where('id','!=',$id)->where($column,'=',$var)->first();
         return (!empty($cek) ? false : true);
@@ -42,7 +42,7 @@ class FinanceController extends Controller
         $start = $request->start;
         $page  = $start +1;
         $search = $request->search['value'];
-        
+
         $country = Finance::select('finance.*', 'komponen_biaya.name');
         $country->leftJoin('komponen_biaya','komponen_biaya.id','finance.komponen_biaya_id');
         if(array_key_exists($request->order[0]['column'], $this->original_column)){
@@ -57,9 +57,9 @@ class FinanceController extends Controller
           });
         }
         $totalData = $country->get()->count();
-  
+
         $totalFiltered = $country->get()->count();
-  
+
         $country->limit($limit);
         $country->offset($start);
         $data = $country->get();
@@ -67,13 +67,17 @@ class FinanceController extends Controller
         {
           $enc_id = $this->safe_encode(Crypt::encryptString($result->id));
           $action = "";
-         
+
           $action.="";
           $action.="<div class='btn-group'>";
-          $action.='<a href="'.route('transaksi.finance.ubah',$enc_id).'" class="btn btn-warning btn-xs icon-btn md-btn-flat product-tooltip" title="Edit"><i class="fa fa-pencil"></i> Edit</a>&nbsp;';
-          $action.='<a href="#" onclick="deleteNegara(this,\''.$enc_id.'\')" class="btn btn-danger btn-xs icon-btn md-btn-flat product-tooltip" title="Hapus"><i class="fa fa-times"></i> Hapus</a>&nbsp;';
+          if($request->user()->can('transaksi.finance.ubah')) {
+                $action.='<a href="'.route('transaksi.finance.ubah',$enc_id).'" class="btn btn-warning btn-xs icon-btn md-btn-flat product-tooltip" title="Edit"><i class="fa fa-pencil"></i> Edit</a>&nbsp;';
+          }
+          if($request->user()->can('transaksi.finance.delete')) {
+            $action.='<a href="#" onclick="deleteNegara(this,\''.$enc_id.'\')" class="btn btn-danger btn-xs icon-btn md-btn-flat product-tooltip" title="Hapus"><i class="fa fa-times"></i> Hapus</a>&nbsp;';
+          }
           $action.="</div>";
-  
+
           if($result->kategori == 0){
             $kategori = 'Debit';
           }else if($result->kategori == 1){
@@ -86,37 +90,32 @@ class FinanceController extends Controller
           $result->kategori       = $kategori;
           $result->action         = $action;
         }
-  
-        $json_data = array(
-          "draw"            => intval($request->input('draw')),  
-          "recordsTotal"    => intval($totalData),
-          "recordsFiltered" => intval($totalFiltered),
-          "data"            => $data
-        );
-        // if($request->user()->can('negara.index')) {
-        //   $json_data = array(
-        //     "draw"            => intval($request->input('draw')),  
-        //     "recordsTotal"    => intval($totalData),
-        //     "recordsFiltered" => intval($totalFiltered),
-        //     "data"            => $data
-        //   );
-        // }else{
-        //   $json_data = array(
-        //     "draw"            => intval($request->input('draw')),  
-        //     "recordsTotal"    => 0,
-        //     "recordsFiltered" => 0,
-        //     "data"            => []
-        //   );
-           
-        // }    
-        return json_encode($json_data); 
-      }
 
-      public function tambah(){
+
+        if($request->user()->can('transaksi.finance.index')) {
+            $json_data = array(
+                "draw"            => intval($request->input('draw')),
+                "recordsTotal"    => intval($totalData),
+                "recordsFiltered" => intval($totalFiltered),
+                "data"            => $data
+            );
+        }else{
+            $json_data = array(
+                "draw"            => intval($request->input('draw')),
+                "recordsTotal"    => 0,
+                "recordsFiltered" => 0,
+                "data"            => []
+            );
+
+        }
+        return json_encode($json_data);
+    }
+
+    public function tambah(){
         return view('backend/pembayaran/finance/form');
-      }
+    }
 
-      public function list_data(Request $request){
+    public function list_data(Request $request){
         $total = $request->total;
 
         echo "
@@ -150,7 +149,7 @@ class FinanceController extends Controller
                         $('#nominal_".$total."').val('');
                         return false;
                     }
-            
+
                     if(value.charAt(0) > 0){
                         $('#nominal_".$total."').val(getprice(nilai));
                     }else{
@@ -220,7 +219,7 @@ class FinanceController extends Controller
               ]
           ]);
         }
-        
+
       }else{
         $komponen = KomponenBiaya::find($request->komponen);
         $result = new Finance;
@@ -272,7 +271,7 @@ class FinanceController extends Controller
         $komponen = KomponenBiaya::find($finance->komponen_biaya_id);
         $finance->tgl_transaksi = date('d-m-Y', strtotime($finance->tgl_transaksi));
         $finance->total = number_format($finance->total,0,',','.');
-        
+
         $detail = FinanceDetail::where('finance_id', $finance->id)->get();
         $total_detail = FinanceDetail::where('finance_id', $finance->id)->count();
         foreach ($detail as $key => $value) {
@@ -281,7 +280,7 @@ class FinanceController extends Controller
       }
       return view('backend/pembayaran/finance/form', compact('enc_id','finance','detail','total_detail','komponen'));
     }
-    
+
     public function hapus(Request $requet, $enc_id){
       $dec_id   = $this->safe_decode(Crypt::decryptString($enc_id));
       $finance = Finance::find($dec_id);
